incremental changes 04/04/2024 4

+++ scripts/prep/aqm/exevs_aqm_aqm_grid2grid_prep.sh
@@ -136,7 +136,7 @@ if [ -d ${log_dir} ]; then
       done
   fi
 fi
-else
+if [ 1 -eq 2 ]; then   ## keep for future one email format
     if [ ${SENDMAIL} = "YES" ]; then
         export subject="${MODELNAME} ${VARID} NC Output Missing for EVS ${COMPONENT}"
         echo "WARNING: No ${MODELNAME} ${VARID} NC output was avaiable valid ${VDATE}" > mailmsg
